+++ src/grid/GridView.php
@@ -314,8 +314,11 @@ class GridView extends BaseListView
 
         $this->initColumns();
 
-        if ((Yii::$app->useJquery ?? false) && !$this->clientScript instanceof ClientScriptInterface) {
-            $this->clientScript ??= ['class' => GridViewJqueryClientScript::class];
+        if ($this->clientScript === null && (Yii::$app->useJquery ?? false)) {
+            $this->clientScript = ['class' => GridViewJqueryClientScript::class];
+        }
+
+        if ($this->clientScript !== null && !$this->clientScript instanceof ClientScriptInterface) {
             $this->clientScript = Yii::createObject($this->clientScript);
         }
     }