+++ src/validators/RequiredValidator.php
@@ -72,8 +72,11 @@ class RequiredValidator extends Validator
             ? Yii::t('yii', '{attribute} cannot be blank.')
             : Yii::t('yii', '{attribute} must be "{requiredValue}".');
 
-        if ((Yii::$app->useJquery ?? false) && !$this->clientScript instanceof ClientValidatorScriptInterface) {
-            $this->clientScript ??= ['class' => RequiredValidatorJqueryClientScript::class];
+        if ($this->clientScript === null && (Yii::$app->useJquery ?? false)) {
+            $this->clientScript = ['class' => RequiredValidatorJqueryClientScript::class];
+        }
+
+        if ($this->clientScript !== null && !$this->clientScript instanceof ClientValidatorScriptInterface) {
             $this->clientScript = Yii::createObject($this->clientScript);
         }
     }